llm_api_server/requirements.txt
fastapi
uvicorn
pydantic
anthropic
google-generativeai
Pillow